Update scripts
- rename trap-msvc-libs.py to trap_msvc_libs.py
- move generator functionality into gen_scripts.py
- update .gitignore and appveyor.yml
