+++ scripts/gen_scripts.py
@@ -1,0 +1,99 @@
+#
+# This file is part of selfrando.
+# Copyright (c) 2015-2017 Immunant Inc.
+# For license information, see the LICENSE file
+# included with selfrando.
+#
+
+from trap_msvc_libs import *
+
+def set_env_vars():
+    cygwin_lines = ["#!/bin/sh","# set env. variables"]
+    pshell_lines = []
+
+    # MSVC_LINKER
+    link_exe = get_path_to_link_exe()
+    pshell_lines.append("$env:MSVC_LINKER_PATH=\"{}\"".format(link_exe))
+
+    link_exe = link_exe.replace("\\", "/") # convert to posix syntax
+    cygwin_lines.append("export MSVC_LINKER_PATH=\"%s\"" %
+        os.path.dirname(link_exe))
+
+    def cygwinify(path):
+        return "/" + path.replace(":", "").replace("\\", "/")
+
+    # PATH
+    scpt_path = os.path.dirname(os.path.join(os.getcwd(), __file__))
+    exes_path = os.path.join(scpt_path, os.pardir, "Release")
+    exes_path = os.path.abspath(exes_path)
+    if os.path.exists(exes_path) and os.path.isdir(exes_path):
+        pshell_lines.append("$env:PATH=\"{}\";$env:PATH".format(exes_path))
+        cygwin_lines.append("export PATH=\"%s\":$PATH" % cygwinify(exes_path))
+    else:
+        exes_path = os.path.join(scpt_path, os.pardir, "Debug")
+        exes_path = os.path.abspath(exes_path)
+        assert os.path.exists(exes_path) and os.path.isdir(exes_path)
+        pshell_lines.append("$env:PATH=\"{}\";$env:PATH".format(exes_path))
+        cygwin_lines.append("export PATH=\"%s\":$PATH" % cygwinify(exes_path))
+
+    # LIB and LIBPATH
+    platform_name = get_platform_name()
+    libs_path = os.path.join(scpt_path, os.pardir, "TrappedMSVCLibs", platform_name)
+    libs_path = os.path.abspath(libs_path)
+    if not os.path.exists(libs_path):
+        os.makedirs(libs_path)
+    else:
+        assert os.path.isdir(libs_path)
+
+    platform_subdir = 'x64' if platform_name == 'x64' else ''
+    randolib_path = os.path.join(scpt_path, os.pardir, platform_subdir, "Release")
+    randolib_path = os.path.abspath(randolib_path)
+    randolib_file_path = os.path.join(randolib_path, "RandoLib.lib")
+    if not os.path.exists(randolib_file_path) or not os.path.isfile(randolib_file_path):
+        randolib_path = os.path.join(scpt_path, os.pardir, platform_subdir, "Debug")
+        randolib_path = os.path.abspath(randolib_path)
+        randolib_file_path = os.path.join(randolib_path, "RandoLib.lib")
+        assert os.path.exists(randolib_file_path) and os.path.isfile(randolib_file_path), \
+               "Invalid RandoLib.lib location: %s" % randolib_path
+
+    pshell_lines.append("$env:LIB=\"{}\";\"{}\";$env:LIB".
+        format(randolib_path, libs_path))
+    cygwin_lines.append("export LIB=\"%s\"\\;\"%s\"\\;$LIB" %
+        (randolib_path, libs_path))
+
+    pshell_lines.append("$env:LIBPATH=\"{}\";\"{}\";$env:LIBPATH".
+        format(randolib_path, libs_path))
+    cygwin_lines.append("export LIBPATH=\"%s\"\\;\"%s\";$LIBPATH" %
+        (randolib_path, libs_path))
+
+    cygwin_outpath = os.path.abspath(os.path.join(scpt_path, "set-buildvars-cygwin-%s.sh" % platform_name))
+    with open(cygwin_outpath, "w") as fh:
+        fh.write("\n".join(cygwin_lines))
+    os.chmod(cygwin_outpath, 0o755)
+
+    pshell_outpath = os.path.abspath(os.path.join(scpt_path, "set-buildvars.ps1"))
+    with open(pshell_outpath, "w") as fh:
+        fh.write("\n".join(pshell_lines))
+
+    print "To set build environment variables:"
+    print " # . {}".format(os.path.basename(cygwin_outpath))
+    print " > . .\{}".format(os.path.basename(pshell_outpath))
+
+
+if __name__ == '__main__':
+
+    # locate the files we need
+    trap_lib_exe = get_exe_path()
+    input_libs = get_libs_from_env()
+
+    # make sure the output directory exists
+    scpt_path = os.path.dirname(os.path.join(os.getcwd(), __file__))
+    platform_name = get_platform_name()
+    out_path = os.path.join(scpt_path, os.pardir, 'TrappedMSVCLibs', platform_name)
+    if not os.path.isdir(out_path):
+        os.makedirs(out_path)
+        print 'Created output directory %s...' % out_path
+
+    # trap_msvc_libs(input_libs)
+
+    set_env_vars()
